Cleaning up files and rehoming things
