+++ Assets/Scripts/Incidents/DelayedIncidentContext.cs
@@ -1,0 +1,14 @@
+﻿namespace Game.Incidents
+{
+	public class DelayedIncidentContext
+	{
+		public IIncidentContext incidentContext;
+		public int delayCounter;
+
+		public DelayedIncidentContext(IIncidentContext context, int delay)
+		{
+			incidentContext = context;
+			delayCounter = delay;
+		}
+	}
+}
