+++ Assets/Scripts/Incidents/DelayedIncidentContext.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e9d65ad594f970a4a8bebf48b3280811
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
