+++ Assets/Scripts/Incidents/FactionContext.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c2a31ca3ab3e18745a96c85a6ac050d5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
