+++ Assets/Scripts/Incidents/Incident.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e6f0a9565281de046ae406b23cd5d8af
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
