+++ Assets/Scripts/Incidents/IncidentActions/IIncidentAction.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using Game.Factions;
+using UnityEngine;
+
+namespace Game.Incidents
+{
+	public interface IIncidentAction
+	{
+		bool VerifyAction(IIncidentContext context, Func<int, IIncidentActionField> delayedAction);
+		void PerformAction(IIncidentContext context);
+		void UpdateEditor();
+	}
+}
