+++ Assets/Scripts/Incidents/IncidentActions/IncidentAction.cs
@@ -1,18 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using Game.Factions;
-using Sirenix.OdinInspector;
-using UnityEngine;
 
 namespace Game.Incidents
 {
-	public interface IIncidentAction
-	{
-		bool VerifyAction(IIncidentContext context, Func<int, IIncidentActionField> delayedAction);
-		void PerformAction(IIncidentContext context);
-		void UpdateEditor();
-	}
-
 	abstract public class IncidentAction<T> : IIncidentAction where T : IIncidentContext
 	{
 		private IIncidentContext context;
@@ -53,34 +42,5 @@ namespace Game.Incidents
 
 		abstract public void UpdateEditor();
 		abstract protected bool VerifyContextActionFields(IIncidentContext context, Func<int, IIncidentActionField> delayedCalculateAction);
-	}
-
-	[Serializable]
-	public class TestFactionIncidentAction : IncidentAction<FactionContext>
-	{
-		[HideReferenceObjectPicker]
-		public IncidentContextActionField<FactionContext> requiredFaction;
-
-		override public void PerformAction(IIncidentContext context)
-		{
-			PerformDebugAction();
-		}
-
-		protected override bool VerifyContextActionFields(IIncidentContext context, Func<int, IIncidentActionField> delayedCalculateAction)
-		{
-			var succeeded = requiredFaction.CalculateField(context, delayedCalculateAction);
-			return succeeded;
-		}
-
-		private void PerformDebugAction()
-		{
-			OutputLogger.Log("Faction Debug Action Performed!");
-		}
-
-		public override void UpdateEditor()
-		{
-			requiredFaction = new IncidentContextActionField<FactionContext>(ContextType);
-			requiredFaction.criteria = new List<IIncidentCriteria>();
-		}
 	}
 }
