+++ Assets/Scripts/Incidents/IncidentActions/IncidentAction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 491ee1bd694b7af479c2b887c3c50cd1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
