+++ Assets/Scripts/Incidents/IncidentContextActionField.cs
@@ -7,18 +7,6 @@ using UnityEngine;
 
 namespace Game.Incidents
 {
-	public interface IIncidentActionField 
-	{
-		int ActionFieldID { get; set; }
-		string ActionFieldIDString { get; }
-		string NameID { get; set; }
-
-		Type ContextType { get; }
-
-		bool CalculateField(IIncidentContext context, Func<int, IIncidentActionField> delayedCalculateAction);
-		IIncidentContextProvider GetFieldValue();
-	}
-
 	public enum ActionFieldRetrievalMethod { Criteria, From_Previous, Random };
 
 	[Serializable]
