+++ Assets/Scripts/Incidents/IncidentCriteria.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 36cfffd9871e59f44b0df85bd49c94ed
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
