+++ Assets/Scripts/Incidents/IncidentCriteriaContainer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1d5dd9ea68cdb1240bb58b5f556160f1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
