+++ Assets/Scripts/Incidents/IncidentReport.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4270f53ff8b422e45a2145244ab58c4e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
